cdp-agentkit-core/python/pyproject.toml
[tool.poetry]
name = "cdp-agentkit-core"
version = "0.0.11"
description = "CDP Agentkit core primitives"
authors = ["John Peterson <john.peterson@coinbase.com>"]
readme = "README.md"
license = "Apache-2.0"
keywords = ["coinbase", "sdk", "crypto", "cdp", "agentkit", "ai", "agent"]
packages = [{ include = "cdp_agentkit_core" }]

[tool.poetry.dependencies]
python = "^3.10"
cdp-sdk = "^0.15.0"
pydantic = "^2.0"
web3 = "^7.6.0"
nilql = "^0.0.0a9"
ecdsa = "^0.19.0"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
jsonschema = "^4.23.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.1"
mypy = "^1.13.0"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
sphinx = "^8.0.2"
sphinx-autobuild = "^2024.9.19"
sphinxcontrib-napoleon = "^0.7"
sphinx-autodoc-typehints = "^2.4.4"
myst-parser = "^4.0.0"
ruff-lsp = "^0.0.58"
python-lsp-server = "^1.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = ["./build/**", "./dist/**", "./docs/**"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "D", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["D213", "D203", "D100", "D104", "D107", "E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["cdp_agentkit_core, cdp"]
